+++ src/core/border.ts
@@ -1,7 +1,7 @@
 import { borderRadius, borderSpacing, borderWidth } from "@/defaults/variables";
 import { standardPreset } from "@/defaults/variants/preset";
 import { createValues } from "@/helpers/create-values";
-import { Utilities } from "@/interfaces";
+import type { Utilities } from "@/interfaces";
 
 export const border: Utilities = {
   "border-collapse": {