+++ src/core/box-model.ts
@@ -1,7 +1,13 @@
-import { dimension, height, margin, padding, width } from "@/defaults/variables";
+import {
+  dimension,
+  height,
+  margin,
+  padding,
+  width,
+} from "@/defaults/variables";
 import { standardPreset } from "@/defaults/variants/preset";
 import { createValues } from "@/helpers/create-values";
-import { Utilities } from "@/interfaces";
+import type { Utilities } from "@/interfaces";
 
 export const boxModel: Utilities = {
   "box-sizing": {