+++ src/core/effect.ts
@@ -9,7 +9,7 @@ import {
   grayscale,
 } from "@/defaults/variables";
 import { standardPreset } from "@/defaults/variants/preset";
-import { Utilities } from "@/interfaces";
+import type { Utilities } from "@/interfaces";
 
 export const effect: Utilities = {
   "backdrop-blur": {