+++ src/core/interactivity.ts
@@ -1,7 +1,7 @@
 import { margin, padding } from "@/defaults/variables";
 import { standardPreset } from "@/defaults/variants/preset";
 import { createValues } from "@/helpers/create-values";
-import { Utilities } from "@/interfaces";
+import type { Utilities } from "@/interfaces";
 
 export const interactivity: Utilities = {
   appearance: {