+++ src/defaults/variables.ts
@@ -61,7 +61,8 @@ export const opacity = 0.1;
 export const bottomLeftTopRight = 0.25;
 
 // font
-export const fontCharter = 'Charter, "Bitstream Charter", "Sitka Text", Cambria, serif';
+export const fontCharter =
+  'Charter, "Bitstream Charter", "Sitka Text", Cambria, serif';
 export const fontMono = '"Nimbus Mono PS", "Courier New", monospace';
 export const fontSystem = "system-ui, sans-serif";