+++ src/helpers/create-values.ts
@@ -5,7 +5,12 @@ type Values = {
   max?: number;
 };
 
-export const createValues = ({ base, unit = "", min = 0, max = 100 }: Values): Record<string, string> => {
+export const createValues = ({
+  base,
+  unit = "",
+  min = 0,
+  max = 100,
+}: Values): Record<string, string> => {
   const values: Record<string, string> = {};
 
   for (let i = min; i <= max; i++) {